[Server] admin page chef 삭제 accessToken 및 refreshToken 유효성 검사

+++ server/controllers/admin/chef.js
@@ -63,15 +63,38 @@ module.exports = {
     const findChefId = await chef.findOne({ id: req.body.id });
     const userId = findChefId.dataValues.chUserId;
 
-    await review.destroy({ where: { rvChefId: req.body.id }, force: true });
-    await reservation.destroy({
-      where: { rsChefId: req.body.id },
-      force: true,
-    });
-    await course.destroy({ where: { coChefId: req.body.id }, force: true });
-    await chef.destroy({ where: { id: req.body.id }, force: true });
-    await user.destroy({ where: { id: userId }, force: true });
+    const accessVerify = isAuthorized(req);
 
-    res.status(200).json({ message: 'ok' });
+    if (!accessVerify) {
+      const refreshVerify = refreshAuthorized(req);
+      if (!refreshVerify) {
+        res.status(401).json({ message: 'Send new Login Request' });
+      } else {
+        delete refreshVerify.exp;
+        const accessToken = basicAccessToken(refreshVerify);
+
+        await review.destroy({ where: { rvChefId: req.body.id }, force: true });
+        await reservation.destroy({
+          where: { rsChefId: req.body.id },
+          force: true,
+        });
+        await course.destroy({ where: { coChefId: req.body.id }, force: true });
+        await chef.destroy({ where: { id: req.body.id }, force: true });
+        await user.destroy({ where: { id: userId }, force: true });
+
+        res.status(201).json({ accessToken, message: 'ok' });
+      }
+    } else {
+      await review.destroy({ where: { rvChefId: req.body.id }, force: true });
+      await reservation.destroy({
+        where: { rsChefId: req.body.id },
+        force: true,
+      });
+      await course.destroy({ where: { coChefId: req.body.id }, force: true });
+      await chef.destroy({ where: { id: req.body.id }, force: true });
+      await user.destroy({ where: { id: userId }, force: true });
+
+      res.status(200).json({ message: 'ok' });
+    }
   },
 };
